+++ packages/terra-clinical-result/tests/jest/result-name-header-cell/ResultNameHeaderCell.test.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import ResultNameHeaderCell from '../../../src/resultnameheadercell/ResultNameHeaderCell';
+import ResultNameHeaderCell from '../../../src/result-name-header-cell/ResultNameHeaderCell';
 
 // Snapshot Tests
 describe('ResultNameHeaderCell', () => {
